rule action balance refactor WIP
