+++ src/holon/models/gridconnection.py
@@ -6,6 +6,7 @@ from modelcluster.models import ClusterableModel
 
 from holon.models.actor import Actor
 from holon.models.gridnode import ElectricGridNode, HeatGridNode
+from holon.models.rule_actions.rule_action_model import RuleActionModel
 from holon.models.scenario import Scenario
 
 holon_app = apps.get_app_config("holon")
@@ -28,7 +29,7 @@ class BatteryMode(models.TextChoices):
     PRICE = "PRICE"
 
 
-class GridConnection(PolymorphicModel, ClusterableModel):
+class GridConnection(PolymorphicModel, ClusterableModel, RuleActionModel):
     owner_actor = models.ForeignKey(Actor, on_delete=models.CASCADE)
     capacity_kw = models.FloatField()
     parent_electric = models.ForeignKey(
