+++ src/holon/tests/test_rule_action_add_remove_set_count.py
@@ -3,7 +3,7 @@ import pytest
 
 from holon.models import *
 from holon.models import rule_mapping
-from holon.models.rule_action import RuleActionFactor
+from holon.models.rule_actions.rule_action import RuleActionFactor
 
 
 class RuleMappingTestClass(TestCase):
